+++ src/modules/timepix3/geometry/Config.h
@@ -12,6 +12,7 @@
 #include <common/debug/Trace.h>
 #include <logical_geometry/ESSGeometry.h>
 #include <string>
+#include <sys/types.h>
 #include <vector>
 
 // #undef TRC_LEVEL
@@ -28,7 +29,7 @@ public:
   uint16_t XResolution{0}; /// Resolution along x axis
   uint16_t YResolution{0}; /// Resolution along y axis
   uint16_t parallelThreads{1};
-  int8_t FrequencyHz{14}; /// Frequency of the detector
+  float FrequencyHz{14}; /// Frequency of the detector
 
   uint32_t MaxTimeGapNS{1};
   uint32_t MinEventSizeHits{1};
